requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0

